+++ src/connect.lua
@@ -1,5 +1,7 @@
 local console = require("console")
-local uri = ...
+local errno = require('errno')
+local uri = require('uri')
+local raw_uri = ...
 
 local TIMEOUT_INFINITY = 100 * 365 * 86400
 local opts = {}
@@ -7,10 +9,11 @@ opts.timeout = TIMEOUT_INFINITY
 
 console.on_start(function(self)
     local status, reason
-    status, reason = pcall(console.connect, uri, opts)
+    status, reason = pcall(console.connect, raw_uri, opts)
     if not status then
-        -- `type(reason) == cdata`, so we have to convert it
-        self:print(tostring(reason))
+        self:print(string.format("%s: %s", reason, errno.strerror()))
+        -- Using urilib to exclude password
+        self:print(string.format("uri: %s", uri.format(uri.parse(raw_uri), false)))
         os.exit(1)
     end
     -- We should only set this after we try to connect, because